Translate FText set on c++
Example on how to translate a simple C++ string by calling a reference to an already translated string inside of it.

All the rest of it is pretty much code dependencies from string files.

I do need though to confirm why it's presenting 68 strings to english and only 21 for other languages, this way it seems it will never be completely translated into the actual tool.